some merges, fixed bugs, moved navigation drawer to its own activity and made login swap to it on success
