+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
+
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
 
     <application
@@ -14,6 +15,9 @@
         android:theme="@style/Theme.ScribeWiz"
         tools:targetApi="31">
         <activity
+            android:name=".NavigationActivity"
+            android:exported="false" />
+        <activity
             android:name=".FirebaseUIActivity"
             android:exported="false">
             <meta-data
@@ -22,8 +26,8 @@
         </activity>
         <activity
             android:name=".MainActivity"
-            android:theme="@style/Theme.ScribeWiz.NoActionBar"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/Theme.ScribeWiz.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
