+++ app/src/main/res/layout/fragment_rec.xml
@@ -6,31 +6,31 @@
 
     <TextView
         android:id="@+id/time_recording"
-        android:layout_width="108dp"
-        android:layout_height="48dp"
+        android:layout_width="175dp"
+        android:layout_height="68dp"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
-        android:layout_marginStart="250dp"
-        android:layout_marginTop="187dp"
-        android:layout_marginEnd="51dp"
-        android:layout_marginBottom="368dp"
+        android:layout_marginStart="213dp"
+        android:layout_marginTop="179dp"
+        android:layout_marginEnd="22dp"
+        android:layout_marginBottom="357dp"
         android:text="00:00"
         android:textSize="18dp" />
 
     <Button
         android:id="@+id/record_button"
-        android:layout_width="169dp"
-        android:layout_height="57dp"
+        android:layout_width="wrap_content"
+        android:layout_height="82dp"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
         android:layout_alignParentEnd="true"
         android:layout_alignParentBottom="true"
         android:layout_marginStart="15dp"
         android:layout_marginTop="171dp"
-        android:layout_marginEnd="225dp"
-        android:layout_marginBottom="374dp"
+        android:layout_marginEnd="214dp"
+        android:layout_marginBottom="350dp"
         android:text="start recording" />
